requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
pip==24.0
setuptools==65.5.0
sqlparse==0.5.1
tzdata==2024.1
django-debug-toolbar==4.4.6
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-spectacular==0.27.2
requests
redis
